Fixed issue #126
Summary:

Sasha proposed to use the certificate principal instead of the host name
when establishing new SSH connections.

What I did:

Replaced `ReadKeys()` function in `auth/init.go` with `ReadIdentity()`
which, instead of a simple "key signer" returns a more comprehensive
structure called "Identity"

The structure has the `Cert` field which can be used to obtain "valid
principals".

The first principal is used as an SSH username, instead of the hostname
like before.