DisableWebUI implementation
We had this flag in the configuration forever, but apparently it was
being ignored.

It allows teleport proxy to start without HTTP UI enabled. This is
useful for proxies that strictly proxy and do nothing else.

I ran into this bug when I first time used this flag for Telecast, it
did not work, so I fixed it.