Improved behaviour of --insecure flag
`tsh` would sometimes ignore --insecure flag. To reproduce:

- copy `/var/lib/teleport` from machine A to machine B
- start Teleport on B

Try to connect with `tsh --proxy=localhost --insecure`. It will fail
because it would try to use local key pool (on machine B) which is not
the same as on A. Instead, --insecure should ignore local keys and
simply ignore certificate validation errors.