Moved HTTP client code from lib/web into lib/client
The purpose of this commit was to remove the lib/client dependency of
lib/web.

lib/client must be dependency-free in order to be reusable.

Next step: make the web UI use the same client code as the CLI. This
will remove a ton of duplicate code making Teleport audit surface area
much smaller.