tctl prerp
We have moved the common (OSS/enterprise) tool code (for tctl) into a
separate 'common' module.

But the code isn't accessible because a lot of symbols are lowercased
and not exported.

This simple commit is just a bunch of gorename calls on common tctl
command structs, so they can be reused in Enterprise teleport.