Tiny change to Makefile to make it less prone to errors
Go 1.10 is more multi-threaded. When playing with it, I noticed rare
compiler failures when trying to write to the same .so file

They are caused by `make -j` (parallel make builds) which we're using to
speed up building multiple binaries at the same time.

Adding ./tool/... as a dependency solves this.

Source-commit: gravitational/teleport.e@17e0ce9b971f97900d510ced15b9fdcd77db6f24