Introduce additional phase to CA rotation.
Flaky tests in teleport integration suite uncovered a problem.
It is possible that main cluster rotates certificate authority,
and will try to dial to the remote cluster with new credentials
before the remote cluster could fetch the new CA to trust.

To fix this, phase "update_clients" was split in two phases:

* Init and Update clients

Init phase does nothing on the main cluster except generating
new certificate authorities, that are trusted but not used in the
cluster.

This phase exists to give remote clusters opporunity
to update the list of trusted certificate authorities
of the main cluster, before main cluster reconnects with new clients
in "Update clients" phase.