Always add 'system:authenticated' to impersonation groups.
Looks like impersonated users do not get 'system:authenticated'
as a default. This leads to kubectl to fail for some restricted users,
who don't have 'system:discovery' like permissions added to the roles.

This commit always injects the group if not present in the request.

See https://github.com/kubernetes/kubernetes/issues/43227 for a similar
solution