Support dial back nodes with Trusted Clusters.
Instantiate agent pool (and agent) with a reference to the reverse
tunnel server.

Pass list of principals to agents when initiating a transport dial
request.

The above two changes allow the agent to look up principals in local
site when attempting to connect to a node within a trusted cluster.