Fix event index for session data.
Session Data transfer event missed event index,
effectively setting it to 0.

It is not a mistake by itself, however on Dynamo
it resulted this event overwriting SessionStart
event, resulting in incomplete session
event records.

This is not a perfect fix, as the same problem
could be introduced at any other point,
so something more robust should be added in the future
to prevent similar problems from happening.