Bump single IP limits.
In Teleport IOT use case every connection
to the auth server is coming from Proxy.

On large clusters that triggers internal rate limiting
on the auth server side that is done per IP.

This commit resolves the issue by bumping the limit to 15K.

However this deserves a second look, as not connections are
equal - sometimes having a stricter limits from unknown IPs
makes more sense.