Allow tsh to connect to legacy clusters.
While moving from lemma (NaCl based) to the new internal secret package
(AES-GCM based), Teleport was updated to allow older tsh clients to
connect to newer proxies. Support to allow newer tsh clients to connect
to older proxies was omitted.

To allow newer tsh clients to connect to older proxies, Teleport
attempts to decrypt the response payload using the new secret
package, and if it fails, attempts to use the legacy lemma package.

In addition, the secret key that tsh generates is encoded in the new as
well as older format when submitting the client submits the request to
Teleport.