Fixed unauthenticated user creation.
The currentUserAction endpoint on the Auth Server allows access to a
user "Nop" if no form of client authentication is provided. Since this
user is not associated with a role, it has no logins, and can not obtain
SSH x509 credentials.

To strengthen access controls, currentUserAction only allows locally
authenticated users whose with a name match access.