Update logic for default auth server selection.
When tctl tries to find the address of the Auth Server to connect to,
first look in file configuration. If a list of Auth Servers is provided
on the CLI override what exists in file configuration.

If nothing is provided in either file configuration or the CLI, then use
the default address.