Added support for session streaming.
Added support to stream sessions from a node to the auth server. This
was added to reduce the memory pressure on the auth server when large
session records are uploaded.

The original behavior was to duplicate the tar/gzip archive in memory,
uncompress it, validate the events in the archive, then pass the copy
along for upload (disk or cloud provided object store).

The new behavior is for nodes to stream chunks to the auth server. If
the chunk is being streamed raw, the chunk will go straight into the
archive (where it will be uploaded as it's assembled). If the chunks are
event chunks, the gzip archive will be created in memory and then
written to the tar archive.

Backpressure is applied by only allowing a limited number of gzip
archive to be constructed at a time and a pool of buffers is used to
allow re-use of heap memory.