Make sure all needed credentials are present in kubeconfig.Update
If TLS client key/cert or CA cert are missing, the kubeconfig ends up
generated successfully, but with those fields empty. For a user, this
looks like a successful `tsh login` with `kubectl` not working
afterwards with cryptic x509 errors.

We should always have the necessary fields provided. If not, `tsh login`
should say exactly what was missing.