Fix integration tests inside container
The way that integration tests are called from the root directory `Makefile` versus the build.assets `Makefile` is currently independent, meaning that if you run locally you get verbose output with a race detector, and if you run inside a container you get no verbose output and no race detector, giving a different experience and potentially changing the outcome of the tests. This changes the process so that it just runs `make integration` from the base `Makefile` inside a container. This is similar to the approach for `make test`.