Use host CA to validate proxy cert on tsh login -o id.pam
When exporting an identity file via tsh login, we won't write the CAs
onto ~/.tsh/known_hosts. But this file is used to validate proxy SSH
cert when dialing it later in the code, after login.

This results in a user getting prompted to verify the proxy public key
manually, even though we have the CA cert in memory.

Manually plumb the trusted SSH CAs to client.HostKeyCallback to get it
validated. Fallback to the existing client.HostKeyCallback if it doesn't
work (which "should never happen").

Fixes https://github.com/gravitational/teleport/issues/3910