Hide fips.go from the "go mod" tools
This is a hack. There's a detailed comment in
`tool/modules/fips.go.hide`.

The short version is: "go mod" doesn't like it when a non-existing
stdlib package is imported, even if that import is behind a build tag.
We have to hide this file from "go mod" (by using a different extension)
while still including it in the build when the "fips" tag is set.

This is the best that me and 2 core Go team members could come up with.
Without this, we can't migrate to modules.

Source-commit: gravitational/teleport.e@56b427fa35c48b628e546e2b900c06dbb818eef2