Prevent multiple auth servers to report excess usage (#147)
This commit fixes a concurrency bug that was (indirectly) reported by
Mattermost. When running multiple instances of this enterprise binary,
we ended up reporting usage multiple times as well.

Note that the _reporting_ part was already concurrency safe (as it was
using CompareAndSwap), but the way we wrote new usage to the internal
backend was not.

To fix this issue, this commit adds a locking mechanism to the
enforcement service. When usage is recorded, we add a lock with a TTL
equal to the reporting interval.

Source-commit: gravitational/teleport.e@4c8b55864a35df0b705bd656cd5d0ffaa34ea142
Source-pull-requests: gravitational/teleport.e#147