Fix context leak in events fanout.
This commit fixes #4511.

Fanout watcher.Close method was cancelling
the context, but did was not removing the watcher
from the fanout list.

GRPC server was not releasing memory buffers associated
with the streams after clients disconnects.

Goroutines associated with the GRPC server were closed,
but buffers remained in memory:

https://github.com/gravitational/teleport/issues/4511
https://github.com/grpc/grpc-go/issues/3728#issuecomment-695883580

In Go, child context created with context.WithValue(parent)
references parent context and parent context references
child context back.

When the parent context is closed, it removes the child references,
but the child keeps referencing the parent context.

If the child context is leaked, objects associated
with the parent context are not garbage collected.

GRPC UnaryInterceptor created context.WithValue(ctx, User)
to add a user and passed this context to methods.

WatchEvents GRPC server created services.Fanout.Watcher
and referenced the child context.

Fanout watcher Close method did not remove the watcher
from the fanout buffer list causing the leak.