lib/utils: factor context-insensitive error formatting from UserMessageFromError.
Split UserMessageFromError in two - one with the same name that keeps
the old behavior of formatting the error via debug report if debug flag
has been specified and the other (upon it builds) to actually do error
formatting. Then use the second implementation to forward error messages
through remote channels to avoid being dependent on whether the proxy is
running with the debug flag.