Changed redirect endpoint when session is expired.
When the user does not have a session, if the user tries to access a
proxied application at it's FQDN, Teleport does best effort resolution.

This fix changes the behavior of what happens when the user has a
session but the session is expired. The user was being redirected to the
login page. This fix changes the behavior to by in sync with the
no-session behavior in doing best effort resolution.