document lower case only in app name
The IsDNS1035Label function that is called in lib/service/cfg.go expects this validation to be lower case only (see https://github.com/kubernetes/kubernetes/issues/71140 for the k8s position on this code behavior)

Our docs say upper case characters are allowed, but they currently are not. Updating this doc to reflect behavior.