Added "tsh env" command.
Added "tsh env" command which prints out the proxy and cluster selected
in the users tsh profile.

If run under "eval", this allows a user to pin a proxy and/or cluster to
a particular terminal. This allows users to run "tsh login" again in
another terminal while keeping the current terminal pinned.

Also drops emitting the TELEPORT_SITE environment variable and instead
starts emitting TELEPORT_CLUSTER. TELEPORT_SITE is still read in for
backward compatibility.