Update compatibility guarantee for client binaries (#5293)
Our previous guarantee was: "tsh/tctl are only guaranteed to work with
teleport binaries of the same version".
Our new guarantee is: "tsh/tctl will work with teleport binaries on the
same or the previous major release version".

This new guarantee is identical to our existing guarantee for server
binaries. It's just extended to apply to clients too, and enables safe
upgrades without access interruption.