Adds md5 auth caveat to Self-Hosted Postgres Guide (#5834)
The existing guide says to that several entries need to be added
to the PostgreSQL authentication rules file.

What is non-obvious is that simply appending these rules may not work,
as a higher-priority rule may override them. If this happens, PostgreSQL
may not let the Teleport DB proxy log in as, for example, as PostgreSQL
will expect a password that the proxy doesn't have.

This change adds a caveat to the `pg_hba.conf` configuration warning
that this might happen so the reader doesn't have to spend time
diagnosing login failures.