Skip directories when enumerating keys
Addresses Issue #5774

If there is a spurious directory in a local keystore directory, `tsh`
will fail when enumerating keys. This is caused by the key enumeration
code attempting to read the directory lioke a key file and failing.

This patch makes the `*-db` and `*-kube` key enumeration skip any
directories it encounters in order for the enumeration to complete.