Update ssh-rbac.mdx to include remote_cluster
The resolution to Issue 5992 and executing the command tctl rm rc/<cluster-name> via the proxy was to add the remote_cluster permissions to the Teleport role.  

The remote_cluster resource is not currently documented in our RBAC documentation.  Adding basic notes that the remote_cluster option is present and configurable in a role configuration.